+++ nf_core/modules/modules_json.py
@@ -432,7 +432,7 @@ class ModulesJson:
             to_name += f"-{datetime.datetime.now().strftime('%y%m%d%H%M%S')}"
         shutil.move(str(current_path), local_dir / to_name)
 
-    def unsynced_components(self) -> Tuple[List[str], List[str], dict]:
+    def unsynced_components(self) -> Tuple[List[str], List[str], Dict]:
         """
         Compute the difference between the modules/subworkflows in the directory and the
         modules/subworkflows in the 'modules.json' file. This is done by looking at all
@@ -1119,8 +1119,10 @@ class ModulesJson:
         """
         Sort the modules.json, and write it to file
         """
+        # Sort the modules.json
+        if self.modules_json is None:
+            self.load()
         if self.modules_json is not None:
-            # Sort the modules.json
             self.modules_json["repos"] = nf_core.utils.sort_dictionary(self.modules_json["repos"])
             if run_prettier:
                 dump_json_with_prettier(self.modules_json_path, self.modules_json)